Refactor: extract SongViewerView from SongbookView.swift
Move the song viewer (870 lines) to its own file, reducing
SongbookView.swift from 1220 to 351 lines. Pure file extraction
with no logic changes.

SongViewerView includes: song info, transpose, auto-scroll,
share/PDF export, strum pattern picker, label management,
chord popover, section navigation, and ChordPro rendering.
